+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 GAMSPy CHANGELOG
 ================
 
+GAMSPy 1.7.0
+------------
+- Testing
+  - Enforce the order of tests. Run unit tests first, and model library tests last.
+  - Use spawn method for multiprocessing to avoid possible deadlocks with fork method.
+
 GAMSPy 1.6.0
 ------------
 - General